Define variables per distribution and version
While there is some redundant data, being explicit about which packages
are installed on which distribution and version is far easier to maintain.

Change the include_vars task to search for files in the follow order:
  - distribution version
  - family version
  - distribution
  - family

This is ordered from most specific to least specific. This allows having a set
of variables that applies to all distributions or all families, except for one
specific version that may need an override.